requirements.txt
# django
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Pillow
pillow==10.4.0